support optional set and enum

+++ golalib/testdata/users/users.go
@@ -378,14 +378,14 @@ func (c *UpdatedAt) UnmarshalJSON(data []byte) error {
 // float
 type FloatType struct {
 	_updated bool
-	val      float32
+	val      float64
 }
 
-func (c *FloatType) GetFloatType() float32 {
+func (c *FloatType) GetFloatType() float64 {
 	return c.val
 }
 
-func (c *FloatType) SetFloatType(val float32) bool {
+func (c *FloatType) SetFloatType(val float64) bool {
 	if c.val == val {
 		return false
 	}
@@ -590,7 +590,6 @@ func (c *Sports) SetSports(val []UserSports) bool {
 	c.val = strings.Join(strSlice, ",")
 	return true
 }
-
 func (c *Sports) IsUpdated() bool {
 	return c._updated
 }
@@ -643,7 +642,6 @@ func (c *Sports2) SetSports2(val []UserSports2) bool {
 	c.val = strings.Join(strSlice, ",")
 	return true
 }
-
 func (c *Sports2) IsUpdated() bool {
 	return c._updated
 }
@@ -696,7 +694,6 @@ func (c *SportsNoDefault) SetSportsNoDefault(val []UserSportsNoDefault) bool {
 	c.val = strings.Join(strSlice, ",")
 	return true
 }
-
 func (c *SportsNoDefault) IsUpdated() bool {
 	return c._updated
 }
@@ -733,7 +730,7 @@ func New() *User {
 		Email{},
 		CreatedAt{val: uint(0)},
 		UpdatedAt{val: uint(0)},
-		FloatType{val: float32(0)},
+		FloatType{val: float64(0)},
 		DoubleType{val: float64(0)},
 		Hobby{val: "swimming"},
 		HobbyNoDefault{},
@@ -752,7 +749,7 @@ func NewWithPK(val int) *User {
 		Email{},
 		CreatedAt{val: uint(0)},
 		UpdatedAt{val: uint(0)},
-		FloatType{val: float32(0)},
+		FloatType{val: float64(0)},
 		DoubleType{val: float64(0)},
 		Hobby{val: "swimming"},
 		HobbyNoDefault{},
